Move send commands to send subcommands
