+++ src/commands/get.command.ts
@@ -100,8 +100,6 @@ export class GetCommand extends DownloadCommand {
                 return await this.getTemplate(id);
             case 'fingerprint':
                 return await this.getFingerprint(id);
-            case 'send':
-                return await this.getSend(id, null, options);
             default:
                 return Response.badRequest('Unknown object.');
         }
@@ -462,66 +460,5 @@ export class GetCommand extends DownloadCommand {
         }
         const res = new StringResponse(fingerprint.join('-'));
         return Response.success(res);
-    }
-
-    private async getSendView(id: string): Promise<SendView | SendView[]> {
-        if (Utils.isGuid(id)) {
-            const send = await this.sendService.get(id);
-            if (send != null) {
-                return await send.decrypt();
-            }
-        } else if (id.trim() !== '') {
-            let sends = await this.sendService.getAllDecrypted();
-            sends = this.searchService.searchSends(sends, id);
-            if (sends.length > 1) {
-                return sends;
-            } else if (sends.length > 0) {
-                return sends[0];
-            }
-        }
-    }
-
-    private async getSend(id: string, filter?: (s: SendView) => boolean, options?: program.OptionValues) {
-        let sends = await this.getSendView(id);
-        if (sends == null) {
-            return Response.notFound();
-        }
-
-        const apiUrl = await this.environmentService.apiUrl;
-        filter = filter == null ? (s: SendView) => true : filter;
-        let selector = async (s: SendView): Promise<Response> => Response.success(new SendResponse(s, apiUrl));
-        if (options.text != null) {
-            filter = s => {
-                return filter(s) && s.text != null;
-            };
-            selector = async s => {
-                // Write to stdout and response success so we get the text string only to stdout
-                process.stdout.write(s.text.text);
-                return Response.success();
-            };
-        }
-        if (options.file != null) {
-            filter = s => {
-                return filter(s) && s.file != null && s.file.url != null;
-            };
-            selector = async s => await this.saveAttachmentToFile(s.file.url, s.cryptoKey, s.file.fileName, options.output);
-        }
-
-        if (Array.isArray(sends)) {
-            if (filter != null) {
-                sends = sends.filter(filter);
-            }
-            if (sends.length > 1) {
-                return Response.multipleResults(sends.map(s => s.id));
-            }
-            if (sends.length > 0) {
-                return selector(sends[0]);
-            }
-            else {
-                return Response.notFound();
-            }
-        }
-
-        return selector(sends);
     }
 }
